+++ doc/v1.2.2/changelog.md
@@ -3,3 +3,4 @@ Find a list of house keeping changes below:
 - Remove dependency to electron-is-dev required package
 - Remove unused variable from configuration.js
 - Remove electron-is-dev package
+- Move default configuration to configuration.js